refactored all test to web helper, web helper is a module
