+++ spec/features/add_new_listing_spec.rb
@@ -1,4 +1,5 @@
 feature 'Add new listings' do
+  include Helpers
   scenario 'page should allow new listings to be created' do
     sign_up
     visit '/spaces/new'
@@ -29,9 +30,7 @@ feature 'Add new listings' do
     sign_up
     create_space_one
     sign_up(email: "a@a.com", password: '1234')
-    click_button "View Space"
-    expect(page).to have_button "Request booking"
-    click_button "Request booking"
+    request_booking
     expect(page).to have_content "Your request to book Space Number One is pending."
   end
 end
