+++ spec/features/approval_spec.rb
@@ -1,14 +1,13 @@
 feature 'Approving requests as a host:' do
-
+include Helpers
  scenario 'Space is removed from list of available spaces when approved' do
    sign_up
    create_space_one
    click_button 'Log Out'
-   fill_in :email, with: 'a@a.com'
-   fill_in :password, with: 'password'
-   click_button "Sign Up"
-   click_button 'View Space'
-   click_button 'Request booking'
+   sign_up_again_and_again
+  #  click_button 'View Space'
+  #  click_button 'Request booking'
+  request_booking
    click_button 'Log Out'
    click_link 'Log In'
    sign_in
@@ -22,11 +21,8 @@ feature 'Approving requests as a host:' do
    sign_up
    create_space_one
    click_button 'Log Out'
-   fill_in :email, with: 'a@a.com'
-   fill_in :password, with: 'password'
-   click_button "Sign Up"
-   click_button 'View Space'
-   click_button 'Request booking'
+   sign_up_again_and_again
+   request_booking
    click_button 'Log Out'
    click_link 'Log In'
    sign_in
@@ -43,17 +39,11 @@ feature 'Approving requests as a host:' do
     sign_up
     create_space_one
     click_button 'Log Out'
-    fill_in :email, with: 'a@a.com'
-    fill_in :password, with: 'password'
-    click_button "Sign Up"
-    click_button 'View Space'
-    click_button 'Request booking'
+    sign_up_again_and_again
+    request_booking
     click_button 'Log Out'
-    fill_in :email, with: 'b@b.com'
-    fill_in :password, with: 'password'
-    click_button "Sign Up"
-    click_button 'View Space'
-    click_button 'Request booking'
+    sign_up_again_and_again(email: 'b@b.com', password: 'password')
+    request_booking
     click_button 'Log Out'
     click_link 'Log In'
     sign_in
@@ -63,10 +53,9 @@ feature 'Approving requests as a host:' do
     expect(page).to have_content "a@a.com requested this"
     end
     within 'a#pending-group' do
-    expect(page).to have_content "b@b.com requested this"
+      expect(page).to have_content "b@b.com requested this"
     end
     expect(Booking.first.status).to eq "Approved"
-
   end
 
 end
